Add handling of error resolution during TVDB update.
For NoMatchFound, get the ChosenName from the ErrorLog, and try it. If it works, mark error as resolved. Trick here was also to use that updated name in the step below, where we check the TVDB name against the TiVo name. I guess I should probably handle the case where the ChosenName works, but isn't an exact match, but ptth.

For Mismatch, just allow the program to continue because it found a match, but also update the "SeriesTitle" field to the ChosenName if it's different. Had to add the new TiVoName column which will hold onto what TiVo is looking for. Had to make replacements places to make the old TiVoUpdater work with that.